beginning optimized solution
this approach attempts to find a potential “center” of a palindrome,
which is either odd (if s[i-1] == s[i+1]) or even (if s[i] == s[i+1]),
and then passes that potential center to a TryEven or TryOdd method.

TryEven and TryOdd methods will utilize a while loop to keep expanding
the candidate palindrome out from i until the mirroring characters no
longer equal; they then return the longest substring for which
mirroring characters did equal.